LPS-72269 Split each table into their own upgrade method
